Move several functions from main.py to buildfunctions.py
